Udownloader/requirements.txt
Django==2.1.2
djangorestframework==3.11.1
pytube3==9.6.4
pytz==2020.1
typing-extensions==3.7.4.3
